# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Security Robot RL API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/environment/definitions/{environment_id}/state": {
      "get": {
        "tags": [
          "environment"
        ],
        "summary": "Get Environment State",
        "operationId": "get_environment_state_api_v1_environment_definitions__environment_id__state_get",
        "parameters": [
          {
            "name": "environment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Environment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/EnvironmentState"
                  },
                  "title": "Response Get Environment State Api V1 Environment Definitions  Environment Id  State Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/environment/sessions": {
      "post": {
        "tags": [
          "environment"
        ],
        "summary": "Create Environment Session",
        "operationId": "create_environment_session_api_v1_environment_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnvironmentSessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/EnvironmentSessionState"
                  },
                  "type": "object",
                  "title": "Response Create Environment Session Api V1 Environment Sessions Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/environment/sessions/{session_id}/reset": {
      "post": {
        "tags": [
          "environment"
        ],
        "summary": "Reset Environment Session",
        "operationId": "reset_environment_session_api_v1_environment_sessions__session_id__reset_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/EnvironmentSessionResetRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/EnvironmentSessionState"
                  },
                  "title": "Response Reset Environment Session Api V1 Environment Sessions  Session Id  Reset Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/environment/sessions/{session_id}/action": {
      "post": {
        "tags": [
          "environment"
        ],
        "summary": "Perform Environment Action",
        "operationId": "perform_environment_action_api_v1_environment_sessions__session_id__action_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnvironmentActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/EnvironmentStepResponse"
                  },
                  "title": "Response Perform Environment Action Api V1 Environment Sessions  Session Id  Action Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/environment/sessions/{session_id}": {
      "delete": {
        "tags": [
          "environment"
        ],
        "summary": "Close Environment Session",
        "operationId": "close_environment_session_api_v1_environment_sessions__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/files/": {
      "post": {
        "tags": [
          "files"
        ],
        "summary": "Upload File",
        "description": "Persist an uploaded file and return its metadata.",
        "operationId": "upload_file_api_v1_files__post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_v1_files__post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/files/list": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "List Files",
        "description": "Return paginated file metadata entries.",
        "operationId": "list_files_api_v1_files_list_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-indexed)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-indexed)"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/files/{file_id}": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Get File Metadata",
        "description": "Retrieve metadata for a single file.",
        "operationId": "get_file_metadata_api_v1_files__file_id__get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileMetadataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "files"
        ],
        "summary": "Delete File",
        "description": "Delete a stored file and its metadata.",
        "operationId": "delete_file_api_v1_files__file_id__delete",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/files/{file_id}/download": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Download File",
        "description": "Stream the stored binary file back to the client.",
        "operationId": "download_file_api_v1_files__file_id__download_get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_file_api_v1_files__post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "file_type": {
            "type": "string",
            "title": "File Type",
            "description": "Logical type of the file (model, log, config, etc.)"
          },
          "training_job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Job Id",
            "description": "Associated training job identifier"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Optional file description"
          },
          "metadata": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Arbitrary JSON metadata as a string"
          }
        },
        "type": "object",
        "required": [
          "file",
          "file_type"
        ],
        "title": "Body_upload_file_api_v1_files__post"
      },
      "EnvironmentActionRequest": {
        "properties": {
          "action": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Action",
            "description": "Discrete action index"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "EnvironmentActionRequest",
        "description": "Request payload for performing an action in an environment session."
      },
...
      "EnvironmentSessionCreate": {
        "properties": {
          "environment_id": {
            "type": "string",
            "minLength": 1,
            "title": "Environment Id"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config",
            "description": "Environment overrides"
          }
        },
        "type": "object",
        "required": [
          "environment_id"
        ],
        "title": "EnvironmentSessionCreate",
        "description": "Request schema for creating an interactive environment session."
      },
      "EnvironmentSessionResetRequest": {
        "properties": {
          "seed": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          }
        },
        "type": "object",
        "title": "EnvironmentSessionResetRequest",
        "description": "Request payload for resetting an environment session."
      },
      "EnvironmentSessionState": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "environment_id": {
            "type": "string",
            "title": "Environment Id"
          },
          "state": {
            "$ref": "#/components/schemas/EnvironmentState"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "environment_id",
          "state"
        ],
        "title": "EnvironmentSessionState",
        "description": "State information for an interactive environment session."
      },
      "EnvironmentState": {
        "properties": {
          "environment_id": {
            "type": "string",
            "title": "Environment Id"
          },
          "definition": {
            "$ref": "#/components/schemas/EnvironmentDefinition"
          },
          "observation": {
            "items": {
              "items": {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Observation"
          },
          "robot": {
            "$ref": "#/components/schemas/RobotState"
          },
          "threat_levels": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Threat Levels"
          },
          "obstacles": {
            "items": {
              "items": {
                "type": "boolean"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Obstacles"
          },
          "suspicious_objects": {
            "items": {
              "$ref": "#/components/schemas/SuspiciousObject"
            },
            "type": "array",
            "title": "Suspicious Objects"
          },
          "time_step": {
            "type": "integer",
            "title": "Time Step"
          },
          "coverage_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Ratio"
          },
          "battery_percentage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Battery Percentage"
          },
          "is_charging": {
            "type": "boolean",
            "title": "Is Charging",
            "default": false
          },
          "distance_to_charging_station": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance To Charging Station"
          },
          "charging_station_position": {
            "anyOf": [
              {
                "prefixItems": [
                  {
                    "type": "integer"
                  },
                  {
                    "type": "integer"
                  }
                ],
                "type": "array",
                "maxItems": 2,
                "minItems": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Charging Station Position"
          }
        },
        "type": "object",
        "required": [
          "environment_id",
          "definition",
          "observation",
          "robot",
          "threat_levels",
          "obstacles",
          "suspicious_objects",
          "time_step"
        ],
        "title": "EnvironmentState"
      },
...
      "EnvironmentStepResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "environment_id": {
            "type": "string",
            "title": "Environment Id"
          },
          "state": {
            "$ref": "#/components/schemas/EnvironmentState"
          },
          "reward": {
            "type": "number",
            "title": "Reward"
          },
          "terminated": {
            "type": "boolean",
            "title": "Terminated"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated"
          },
          "info": {
            "additionalProperties": true,
            "type": "object",
            "title": "Info"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "environment_id",
          "state",
          "reward",
          "terminated",
          "truncated"
        ],
        "title": "EnvironmentStepResponse",
        "description": "Response model for a single environment step."
      },
...
      "FileListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileMetadataResponse"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "page_size",
          "files"
        ],
        "title": "FileListResponse",
        "description": "Response schema for file list."
      },
      "FileMetadataResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "training_job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Job Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "original_filename",
          "file_path",
          "file_size",
          "file_type",
          "content_type",
          "training_job_id",
          "description",
          "created_at",
          "updated_at"
        ],
        "title": "FileMetadataResponse",
        "description": "Response schema for file metadata."
      },
      "FileUploadResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "training_job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Job Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "upload_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "original_filename",
          "file_path",
          "file_size",
          "file_type",
          "content_type",
          "training_job_id",
          "description",
          "created_at",
          "updated_at"
        ],
        "title": "FileUploadResponse",
        "description": "Response schema for file uploads."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}